refactor: palette -> palette-editor
