+++ src/components/v1/palette/index.ts
@@ -2,6 +2,7 @@ export * from './color-blindness-simulator';
 export * from './controls';
 export * from './feedback-widget';
 export * from './header-icon-stack';
-export * from './palette';
+export * from './palette-editor';
+export * from './palette-name-input';
 export * from './panels';
 export * from './swatch';
