+++ src/components/v1/palette/palette-editor.tsx
@@ -1,0 +1,53 @@
+import { useDisclosure } from '@chakra-ui/react';
+import { useCallback, useState } from 'react';
+import { Palette } from './palette';
+import { PaletteNameInput } from './palette-name-input';
+
+import { usePaletteState } from '@/contexts';
+import { isDev } from '@/utils';
+import ColorLab from 'lib/color';
+import { CommandPalette } from '../_wip/command-palette';
+import { LogoCredits } from '../_wip/logo-credits';
+import { ColorBlindnessSimulator } from './color-blindness-simulator';
+import { HeaderIconStack } from './header-icon-stack';
+import { ExportPanel } from './panels/export-panel';
+
+export function PaletteEditor() {
+  const { palette } = usePaletteState();
+  const [showCB, setShowCB] = useState(false);
+  const showColorBlindness = useCallback(
+    () => setShowCB(!showCB),
+    [showCB, setShowCB]
+  );
+
+  const { isOpen, onOpen, onClose } = useDisclosure(); // export modal
+
+  const contrast =
+    new ColorLab(palette[0]!).contrast == 'dark' ? 'whiteAlpha' : 'blackAlpha';
+
+  const controlsText =
+    new ColorLab(palette[palette.length - 1]!).contrast == 'dark'
+      ? 'gray.600'
+      : 'gray.300';
+
+  return (
+    <>
+      {/* @TODO: WIP: finish command Palette logic */}
+      {isDev ? <CommandPalette /> : null}
+      <HeaderIconStack
+        palette={palette}
+        openModal={onOpen}
+        showColorBlindness={showColorBlindness}
+        showCB={showCB}
+      />
+      <ExportPanel isOpen={isOpen} onClose={onClose} />
+      {showCB ? (
+        <ColorBlindnessSimulator palette={palette} contrast={contrast} />
+      ) : null}
+      {palette.length && <PaletteNameInput text={controlsText} />}
+      <Palette palette={palette} />
+      {/* @TODO: WIP: Finish LogoCredits */}
+      <LogoCredits />
+    </>
+  );
+}
