+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from 'react';
 
 import type { GetServerSideProps, NextPage } from 'next';
 
-import { BaseLayout, Palette } from '@/components';
+import { BaseLayout, PaletteEditor } from '@/components';
 import { SocialShare } from '@/components/v1/_scaffold/social';
 import { PaletteProvider } from '@/contexts';
 import { getBuildUrl, parsePalette, slugify } from '@/utils';
@@ -45,7 +45,7 @@ const Home: NextPage<Props> = ({ params, paletteName, colorParams }) => {
         <FullScreenLoader color="green" />
       ) : (
         <PaletteProvider paletteName={paletteName!} colorParams={colorParams!}>
-          <Palette />
+          <PaletteEditor />
           <SocialShare twitter pinterest />
         </PaletteProvider>
       )}
